requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docutils==0.20.1
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Kivy==2.3.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
numpy==1.26.4
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.1
